DirectX 10 renderer, and removal of linux platform, then some resource file hate
I know some people are sad about the removal of Linux as a platform, but
it became too hard to try to massage Visual Studio 2015, and then VS2013
was mad about GLEW and GLFW, so I am going to try and write the DX10
renderer from scratch, so anyone who has VS2013 can compile and run.
The test icon is not the final icon (by any means, I can't draw), and
probably won't be done near most likely early April, because we neeeeeed
to focus on engine dev.